refactor: split configuration domains
