+++ src/config.rs
@@ -19,12 +19,15 @@ use crate::i18n::{DjGemLanguage, Language};
 use crate::queue::Repeat;
 use crate::search_source::SearchConfig;
 use crate::streaming::StreamingConfig;
-use crate::t;
 use crate::theme::ThemeConfig;
 
+mod animation;
 mod audio;
 mod recovery;
 mod spotify;
+mod storage;
+mod visual;
+pub use animation::{AnimationsConfig, FPS_DEFAULT, FPS_MAX, FPS_MIN};
 pub use audio::{
     AudioBackend, AudioConfig, AudioRuntimeConfig, LongFormSeekOptimization,
     MPV_CACHE_BACK_DEFAULT, MPV_CACHE_BACK_LEGACY_DEFAULT, MPV_CACHE_DEFAULTS_REVISION,
@@ -32,6 +35,16 @@ pub use audio::{
     MpvAudioRuntimeConfig,
 };
 pub use spotify::SpotifyImportMode;
+pub use storage::{default_cookies_file, default_download_dir, default_recording_dir};
+pub use visual::{AlbumArtQuality, PlayerBarPosition, VideoOverlay};
+
+pub(crate) use storage::config_path;
+#[cfg(test)]
+use storage::{config_for_missing_profile, import_old_from, ytm_dir_under_audio_dir};
+use storage::{
+    external_cookies_warning, import_external_cookies_file, normalize_user_dir,
+    parse_netscape_cookies, validate_external_cookies_file,
+};
 
 /// Clamp range for playback speed (matches the `>`/`<` controls and the settings slider).
 pub const SPEED_MIN: f64 = 0.5;
@@ -120,394 +133,6 @@ pub const RECORDING_MAX_SECONDS_DEFAULT: u32 = 900;
 pub const RECORDING_PAST_TRACKS_MIN: usize = 1;
 pub const RECORDING_PAST_TRACKS_MAX: usize = 50;
 pub const RECORDING_PAST_TRACKS_DEFAULT: usize = 10;
-
-/// UI eye-candy toggles (the **Animations** settings tab). Every field is an
-/// independent on/off; **all default to `false`** so a fresh install behaves exactly like
-/// before (the app's whole identity is "fast and light"). `master` is a global kill-switch:
-/// when it is off, nothing animates regardless of the per-effect flags, and the animation
-/// frame-clock never even wakes (see `App::animation_active`). Grouped under one JSON key
-/// (`"animations"`) and `#[serde(default)]` so older config files forward-migrate cleanly.
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(default)]
-pub struct AnimationsConfig {
-    /// Global enable. Off → the player renders identically to today, zero overhead.
-    pub master: bool,
-    /// Dedicated-Radio-mode override for `master`. `None` inherits `master` (existing
-    /// configs keep behaving as one global switch); the first ✨/`A` toggle taken while in
-    /// Radio mode pins it, after which the two modes animate independently. A scope
-    /// selector, not an effect — deliberately excluded from [`Self::any_effect`]. Resolved
-    /// through [`Self::effective`]; there is no Settings row for it.
-    pub radio_master: Option<bool>,
-    // Element-level effects (restyle existing widgets in place) -----------------
-    /// Shimmer + marquee scroll on the now-playing title line.
-    pub title: bool,
-    /// Pulse the `♥` like-marker when the track is in the library.
-    pub heart: bool,
-    /// Seekbar motion: the sweeping comet on the filled gauge plus smooth sub-second fill
-    /// (the gauge interpolates between mpv's one-per-second reports while the clock runs).
-    pub seekbar: bool,
-    /// Spinning throbber next to "▸ playing" on the status line.
-    pub spinner: bool,
-    /// Faux VU `▁▂▃▅▇` bars on the status line (and a mini VU marker on the queue window's
-    /// now-playing row).
-    pub eq_bars: bool,
-    /// Pulse/glow the transport play-pause glyph.
-    pub controls: bool,
-    /// "Breathing" outer border colour cycle.
-    pub border: bool,
-    // Player one-shots (event feedback, each plays once and the clock re-sleeps) -
-    /// Letter-cascade reveal of the title line when a new track starts.
-    pub track_intro: bool,
-    /// Synced-lyrics polish: the current line breathes and flashes as it becomes current;
-    /// far lines fade with distance.
-    pub lyrics: bool,
-    /// Transient status messages type themselves in with a bright caret head.
-    pub toast: bool,
-    /// A short volume gauge flashes under the transport strip when the volume changes.
-    pub volume_flash: bool,
-    /// A little burst of hearts/sparks around the title when the track is liked.
-    pub like_burst: bool,
-    /// A bright ripple at the seekbar head after a seek.
-    pub seek_flash: bool,
-    // UI-wide effects (Search / Library / Settings / DJ Gem, not just the player) -
-    /// The focused list selection bar breathes gently toward the accent colour.
-    pub selection: bool,
-    /// List rows cascade in top-to-bottom on view/tab switches and new search results.
-    pub stagger: bool,
-    /// Text-input carets blink (search box, filter, playlist names, settings fields, DJ Gem).
-    pub caret: bool,
-    /// The active tab pops with a brief accent wash on view/tab switches.
-    pub tabs: bool,
-    /// Popups and dropdowns materialize with a ~150 ms fade-in instead of appearing at once.
-    pub popup_fade: bool,
-    /// Activity indicators animate: "Searching…" dots, lyrics fetching, DJ Gem "…thinking",
-    /// and a spinner on a running download's `⬇ N%` tag.
-    pub activity: bool,
-    /// The About card twinkles: sparkles around the icon and a gradient sweep on the name.
-    pub about_fx: bool,
-    // Second-wave Now Playing element effects ----------------------------------
-    /// The seekbar's gauge and time label glow briefly as each playback second lands.
-    pub time_glow: bool,
-    /// Tiny sparks twinkle around the seekbar head while playback runs.
-    pub progress_sparkle: bool,
-    /// A short bright comet chases clockwise around the Player view's outer border.
-    pub border_chase: bool,
-    /// A light wave washes across the transport controls when play/pause toggles.
-    pub pause_flash: bool,
-    /// Error status messages shake side-to-side with a decaying oscillation.
-    pub error_shake: bool,
-    // Filler-canvas effects (drawn only in blank zones) ------------------------
-    /// Matrix-style digital rain in the free zone(s).
-    pub rain: bool,
-    /// Classic spinning ASCII donut.
-    pub donut: bool,
-    /// Decorative (non-audio-reactive) spectrum visualizer.
-    pub visualizer: bool,
-    /// Drifting stars / musical notes.
-    pub starfield: bool,
-    /// DVD-style bouncing logo.
-    pub bounce: bool,
-    /// Occasional diagonal shooting-star streaks.
-    pub comets: bool,
-    /// Sparse drifting snowfall.
-    pub snow: bool,
-    /// Fireflies wandering on smooth glowing paths.
-    pub fireflies: bool,
-    /// Rotating 3-D wireframe cube.
-    pub cube: bool,
-    /// ASCII aquarium: fish swimming both ways plus rising bubbles.
-    pub aquarium: bool,
-    /// Layered ocean waves along the bottom of the free zone.
-    pub waves: bool,
-    /// Periodic firework launches with radial particle bursts.
-    pub fireworks: bool,
-    /// Conway's Game of Life colony, colour-coded by cell age.
-    pub life: bool,
-    /// The classic pipes screensaver growing across the free zone.
-    pub pipes: bool,
-    /// Demoscene plasma colour field over the whole free zone (the heaviest effect).
-    pub plasma: bool,
-    // Behaviour knobs (not effects) -------------------------------------------
-    /// Park the animation tick while the terminal is unfocused (minimized or behind another
-    /// window). Defaults to `true`; opt out to keep animating off-screen. No-op on terminals that
-    /// don't report focus (DECSET ?1004). See [`crate::app::App::animation_active`].
-    pub pause_unfocused: bool,
-    /// Target animation frame rate. Read through [`Self::effective_fps`], which clamps to
-    /// [`FPS_MIN`]..=[`FPS_MAX`] so a hand-edited or corrupt config can't spin the loop or freeze
-    /// it. Lower values trade smoothness for less CPU/battery. Default [`FPS_DEFAULT`].
-    pub fps: u16,
-}
-
-/// Animation frame-rate bounds. The floor keeps motion perceptible; the ceiling caps the
-/// redraw cost. The default matches the long-standing fixed ~30 fps tick.
-pub const FPS_MIN: u16 = 5;
-pub const FPS_MAX: u16 = 60;
-pub const FPS_DEFAULT: u16 = 30;
-
-impl Default for AnimationsConfig {
-    /// All visual effects start **off** (reduced-motion by default); the behaviour knobs default
-    /// to `pause_unfocused: true` and `fps: 30`. A manual impl (rather than `#[derive(Default)]`)
-    /// is required so these aren't `bool`'s `false` / `u16`'s `0` (a `0` fps would clamp to the
-    /// floor and silently override the intended default).
-    fn default() -> Self {
-        Self {
-            master: false,
-            radio_master: None,
-            title: false,
-            heart: false,
-            seekbar: false,
-            spinner: false,
-            eq_bars: false,
-            controls: false,
-            border: false,
-            track_intro: false,
-            lyrics: false,
-            toast: false,
-            volume_flash: false,
-            like_burst: false,
-            seek_flash: false,
-            selection: false,
-            stagger: false,
-            caret: false,
-            tabs: false,
-            popup_fade: false,
-            activity: false,
-            about_fx: false,
-            time_glow: false,
-            progress_sparkle: false,
-            border_chase: false,
-            pause_flash: false,
-            error_shake: false,
-            rain: false,
-            donut: false,
-            visualizer: false,
-            starfield: false,
-            bounce: false,
-            comets: false,
-            snow: false,
-            fireflies: false,
-            cube: false,
-            aquarium: false,
-            waves: false,
-            fireworks: false,
-            life: false,
-            pipes: false,
-            plasma: false,
-            pause_unfocused: true,
-            fps: FPS_DEFAULT,
-        }
-    }
-}
-
-impl AnimationsConfig {
-    /// The frame rate to actually drive the tick at, clamped to a sane range so a bad config
-    /// value (0, or absurdly high) can't busy-spin or stall the animation loop.
-    pub fn effective_fps(&self) -> u16 {
-        self.fps.clamp(FPS_MIN, FPS_MAX)
-    }
-
-    /// Whether any individual effect is enabled (ignores `master`).
-    pub fn any_effect(&self) -> bool {
-        self.title
-            || self.heart
-            || self.seekbar
-            || self.spinner
-            || self.eq_bars
-            || self.controls
-            || self.border
-            || self.track_intro
-            || self.lyrics
-            || self.toast
-            || self.volume_flash
-            || self.like_burst
-            || self.seek_flash
-            || self.selection
-            || self.stagger
-            || self.caret
-            || self.tabs
-            || self.popup_fade
-            || self.activity
-            || self.about_fx
-            || self.time_glow
-            || self.progress_sparkle
-            || self.border_chase
-            || self.pause_flash
-            || self.error_shake
-            || self.any_canvas()
-    }
-
-    /// Whether any filler-canvas effect is enabled — the group `ui::anim::render_canvas`
-    /// dispatches, drawn only into blank zones.
-    pub fn any_canvas(&self) -> bool {
-        self.bounce || self.any_canvas_heavy()
-    }
-
-    /// Canvas effects that repaint enough cells per frame to earn the reduced draw-fps cap
-    /// and DEC synchronized update: every canvas effect except the single-label `bounce`.
-    pub fn any_canvas_heavy(&self) -> bool {
-        self.rain
-            || self.donut
-            || self.visualizer
-            || self.starfield
-            || self.comets
-            || self.snow
-            || self.fireflies
-            || self.cube
-            || self.aquarium
-            || self.waves
-            || self.fireworks
-            || self.life
-            || self.pipes
-            || self.plasma
-    }
-
-    /// Whether animations should actually run: the master switch is on *and* at least one
-    /// effect is enabled. When this is `false`, the per-frame animation clock stays asleep.
-    pub fn active(&self) -> bool {
-        self.master && self.any_effect()
-    }
-
-    /// The config as the render/gating layer should see it in the given mode: `master`
-    /// resolves to the Radio override while dedicated Radio mode is active (`None` =
-    /// inherit). Callers must keep persisting the *stored* config — saving this resolved
-    /// copy would bake the inherit link into the file.
-    pub fn effective(self, radio: bool) -> Self {
-        Self {
-            master: if radio {
-                self.radio_master.unwrap_or(self.master)
-            } else {
-                self.master
-            },
-            ..self
-        }
-    }
-}
-
-/// Source/detail level for remote album art shown inside the terminal. `High` preserves the
-/// historical max-resolution preference and 768px cap; `Original` keeps the fetched source
-/// dimensions intact. Persisted in `config.json`.
-#[derive(Debug, Clone, Copy, Default, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum AlbumArtQuality {
-    Standard,
-    #[default]
-    High,
-    Original,
-}
-
-impl AlbumArtQuality {
-    /// Step through `Standard → High → Original`, wrapping in either direction.
-    pub fn cycled(self, forward: bool) -> Self {
-        match (self, forward) {
-            (Self::Standard, true) => Self::High,
-            (Self::High, true) => Self::Original,
-            (Self::Original, true) => Self::Standard,
-            (Self::Standard, false) => Self::Original,
-            (Self::High, false) => Self::Standard,
-            (Self::Original, false) => Self::High,
-        }
-    }
-
-    /// Short human label for the Playback settings row.
-    pub fn label(self) -> &'static str {
-        match self {
-            Self::Standard => t!("Standard · up to 640 px", "표준 · 최대 640 px"),
-            Self::High => t!("High · up to 768 px", "고화질 · 최대 768 px"),
-            Self::Original => t!("Original source", "원본 화질"),
-        }
-    }
-}
-
-/// Window layout for the external mpv video overlay launched from the player (`v`), cycled
-/// live with `Shift+V` and chosen as the open default in Settings. `Compact` docks a small
-/// ~30% window top-right; `Large` centers a ~50% window; `Fullscreen` fills the screen.
-/// Persisted in `config.json`.
-#[derive(Debug, Clone, Copy, Default, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum VideoOverlay {
-    #[default]
-    Compact,
-    Large,
-    Fullscreen,
-}
-
-impl VideoOverlay {
-    /// Step to the next/previous layout in the `Compact → Large → Fullscreen` cycle.
-    pub fn cycled(self, forward: bool) -> Self {
-        match (self, forward) {
-            (Self::Compact, true) => Self::Large,
-            (Self::Large, true) => Self::Fullscreen,
-            (Self::Fullscreen, true) => Self::Compact,
-            (Self::Compact, false) => Self::Fullscreen,
-            (Self::Large, false) => Self::Compact,
-            (Self::Fullscreen, false) => Self::Large,
-        }
-    }
-
-    /// The next layout (for the forward-cycling `Shift+V` toggle).
-    pub fn toggled(self) -> Self {
-        self.cycled(true)
-    }
-
-    /// Short human label for the status toast.
-    pub fn label(self) -> &'static str {
-        match self {
-            Self::Compact => t!("top-right · 30%", "우상단 · 30%"),
-            Self::Large => t!("center · 50%", "가운데 · 50%"),
-            Self::Fullscreen => t!("fullscreen", "전체화면"),
-        }
-    }
-
-    /// mpv flags for the overlay window. `Compact` docks a borderless top-right ~30% window;
-    /// `Large` a borderless centered ~50% window; `Fullscreen` fills the screen (borderless/
-    /// on-top/autofit are meaningless there, so they're dropped).
-    pub fn mpv_window_args(self) -> Vec<String> {
-        match self {
-            Self::Compact => vec![
-                "--ontop".to_owned(),
-                "--no-border".to_owned(),
-                "--autofit=30%".to_owned(),
-                "--geometry=-20+20".to_owned(),
-            ],
-            Self::Large => vec![
-                "--ontop".to_owned(),
-                "--no-border".to_owned(),
-                "--autofit=50%".to_owned(),
-            ],
-            Self::Fullscreen => vec!["--fullscreen".to_owned()],
-        }
-    }
-}
-
-/// Where the player control block (title / seekbar / transport / status) sits. `Top` is the
-/// legacy layout: the block heads the Player view and other screens carry no player chrome.
-/// `Bottom` (the default) docks the block above the footer on every screen, and the Player
-/// view centers its filler in the space above. Persisted in `config.json`.
-#[derive(Debug, Clone, Copy, Default, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum PlayerBarPosition {
-    Top,
-    #[default]
-    Bottom,
-}
-
-impl PlayerBarPosition {
-    /// The other position (for the Settings `< >` cycle — two states, so both directions agree).
-    pub fn toggled(self) -> Self {
-        match self {
-            Self::Top => Self::Bottom,
-            Self::Bottom => Self::Top,
-        }
-    }
-
-    /// Short human label for the Settings row.
-    pub fn label(self) -> &'static str {
-        match self {
-            Self::Top => t!("Top (classic)", "상단 (클래식)"),
-            Self::Bottom => t!("Bottom (docked)", "하단 (도킹)"),
-        }
-    }
-}
 
 /// Scrobbling accounts (the **Accounts** settings tab) plus shared behavior, grouped under
 /// one JSON key (`"scrobble"`). Every field is `#[serde(default)]` so older config files
@@ -1003,65 +628,6 @@ impl Default for Config {
 }
 
 impl Config {
-    /// Defaults written only for a genuinely new profile. Keep this separate from [`Default`]:
-    /// serde recovery, legacy configs, tests, and non-TUI reset paths all rely on conservative
-    /// defaults that do not opt an existing user into a walkthrough.
-    pub(crate) fn fresh_install() -> Self {
-        Self {
-            beginner_mode: true,
-            beginner_tutorial: BeginnerTutorialProgress::welcome(),
-            ..Self::default()
-        }
-    }
-
-    pub(crate) fn preflight_persistence_recovery()
-    -> Result<(), crate::persist::StartupRecoveryError> {
-        let Some(path) = config_path() else {
-            return Ok(());
-        };
-        crate::persist::preflight_journal_recovery::<Self>(
-            crate::persist::StoreKind::Config,
-            &path,
-            MAX_CONFIG_BYTES,
-        )
-    }
-
-    /// Load config, importing from the old app on first run. Never fails: a missing or
-    /// corrupt file falls back to defaults (+ migration).
-    pub fn load() -> Self {
-        let Some(path) = config_path() else {
-            // No config dir on this platform: migrate from the old app but don't persist.
-            let old = old_config_path();
-            let mut cfg = config_for_missing_profile(old.as_deref());
-            if let Some(old) = old {
-                import_old_from(&old, &mut cfg);
-            }
-            return cfg;
-        };
-        Self::load_from(&path)
-    }
-
-    /// Load from an explicit path (also the test seam). `config.json` is the only
-    /// secret-bearing store (YTM cookie, Gemini key, Last.fm/ListenBrainz tokens), so a
-    /// present-but-unloadable file is set aside *before* the defaults `save()` below can
-    /// overwrite it — matching `safe_fs::load_json_or_default_limited`'s backup-aside policy.
-    fn load_from(path: &std::path::Path) -> Self {
-        recovery::load_from_path(path)
-    }
-
-    /// Persist atomically (write a temp file, then rename over the target).
-    pub fn save(&self) -> std::io::Result<()> {
-        crate::persist::ensure_persistence_writes_allowed()?;
-        let Some(path) = config_path() else {
-            return Ok(()); // no config dir on this platform; nothing to do
-        };
-        self.save_to(&path)
-    }
-
-    fn save_to(&self, path: &std::path::Path) -> std::io::Result<()> {
-        crate::persist::write_config_json(path, self)
-    }
-
     pub fn player_runtime(&self, cookies_file: Option<PathBuf>) -> PlayerRuntimeConfig {
         PlayerRuntimeConfig {
             volume: self.volume,
@@ -1498,219 +1064,6 @@ impl Config {
             other => other,
         }
     }
-}
-
-/// Default location for an optional exported Netscape cookies file.
-///
-/// macOS: `~/Music/yututui/cookies.txt`
-/// Windows: `%USERPROFILE%\Music\yututui\cookies.txt`
-pub fn default_cookies_file() -> Option<PathBuf> {
-    default_ytm_dir().map(|dir| dir.join("cookies.txt"))
-}
-
-fn validate_external_cookies_file(path: &Path) -> std::io::Result<()> {
-    let meta = fs::symlink_metadata(path)?;
-    if meta.file_type().is_symlink() {
-        return Err(std::io::Error::new(
-            std::io::ErrorKind::PermissionDenied,
-            "cookies file must not be a symlink",
-        ));
-    }
-    #[cfg(windows)]
-    {
-        use std::os::windows::fs::MetadataExt;
-        const FILE_ATTRIBUTE_REPARSE_POINT: u32 = 0x0000_0400;
-        if meta.file_attributes() & FILE_ATTRIBUTE_REPARSE_POINT != 0 {
-            return Err(std::io::Error::new(
-                std::io::ErrorKind::PermissionDenied,
-                "cookies file must not be a reparse point",
-            ));
-        }
-    }
-    if !meta.is_file() {
-        return Err(std::io::Error::new(
-            std::io::ErrorKind::InvalidInput,
-            "cookies path is not a regular file",
-        ));
-    }
-    if meta.len() > MAX_COOKIE_BYTES {
-        return Err(std::io::Error::new(
-            std::io::ErrorKind::InvalidData,
-            "cookies file is too large",
-        ));
-    }
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::PermissionsExt;
-        if meta.permissions().mode() & 0o077 != 0 {
-            tracing::warn!(
-                "cookies file is readable outside the owner; importing a private external-tool copy"
-            );
-        }
-    }
-    Ok(())
-}
-
-fn import_external_cookies_file(source: &Path, data_dir: &Path) -> std::io::Result<PathBuf> {
-    validate_external_cookies_file(source)?;
-    let bytes = safe_fs::read_no_symlink_limited(source, MAX_COOKIE_BYTES)?;
-    let target = data_dir.join(EXTERNAL_COOKIES_COPY);
-    safe_fs::write_private_atomic(&target, &bytes)?;
-    Ok(target)
-}
-
-fn external_cookies_warning(error: &std::io::Error) -> String {
-    let reason = match error.kind() {
-        std::io::ErrorKind::NotFound => "file was not found",
-        std::io::ErrorKind::PermissionDenied => "permission or symlink check failed",
-        std::io::ErrorKind::InvalidInput => "path is not a regular file",
-        std::io::ErrorKind::InvalidData => "file is too large or invalid",
-        _ => "file could not be read or imported",
-    };
-    format!(
-        "Cookies file not used for mpv/yt-dlp: {reason}. Use a real non-symlink cookies.txt under 4 MiB."
-    )
-}
-
-/// Default directory for downloaded tracks.
-///
-/// macOS: `~/Music/yututui`
-/// Windows: `%USERPROFILE%\Music\yututui`
-pub fn default_download_dir() -> PathBuf {
-    default_ytm_dir().unwrap_or_else(|| PathBuf::from("yututui"))
-}
-
-/// Normalize a user-supplied directory (env override or stored config): trim surrounding
-/// whitespace, treat a whitespace-only value as "unset" (→ fall through to the default), and
-/// expand a leading `~` / `~/` to the home directory. Without this a value like `~/Music` was
-/// stored/used literally (creating a dir named `~`), and `"   "` created a spaces-named dir.
-fn normalize_user_dir(raw: &str) -> Option<PathBuf> {
-    let trimmed = raw.trim();
-    if trimmed.is_empty() {
-        return None;
-    }
-    if trimmed == "~" {
-        return directories::BaseDirs::new().map(|b| b.home_dir().to_path_buf());
-    }
-    if let Some(rest) = trimmed.strip_prefix("~/")
-        && let Some(base) = directories::BaseDirs::new()
-    {
-        return Some(base.home_dir().join(rest));
-    }
-    Some(PathBuf::from(trimmed))
-}
-
-/// Default directory for saved radio recordings: `<music>/yututui/recordings`.
-pub fn default_recording_dir() -> PathBuf {
-    default_ytm_dir()
-        .map(|d| d.join("recordings"))
-        .unwrap_or_else(|| PathBuf::from("yututui/recordings"))
-}
-
-fn default_ytm_dir() -> Option<PathBuf> {
-    directories::UserDirs::new()
-        .and_then(|u| u.audio_dir().map(std::path::Path::to_path_buf))
-        .map(ytm_dir_under_audio_dir)
-}
-
-fn ytm_dir_under_audio_dir(audio_dir: PathBuf) -> PathBuf {
-    audio_dir.join("yututui")
-}
-
-pub(crate) fn config_path() -> Option<PathBuf> {
-    crate::paths::config_dir().map(|d| d.join("config.json"))
-}
-
-fn old_config_path() -> Option<PathBuf> {
-    // Never import the original app's config while a config-dir override is active (a non-blank
-    // env override — mirroring `paths::config_dir`'s blank-is-unset rule — or the test sandbox):
-    // tests must not read the developer's real `~/.youtube-music-cli`.
-    let overridden = std::env::var("YTM_CONFIG_DIR").is_ok_and(|v| !v.trim().is_empty());
-    if overridden || cfg!(test) {
-        return None;
-    }
-    directories::BaseDirs::new().map(|d| d.home_dir().join(".youtube-music-cli/config.json"))
-}
-
-/// Pick the base for a missing current config. A legacy path that cannot be inspected counts as
-/// present: surprising an existing user with onboarding is worse than conservatively leaving it
-/// off, and the import itself will still use the hardened bounded/no-symlink reader below.
-fn config_for_missing_profile(old: Option<&Path>) -> Config {
-    let legacy_present = old.is_some_and(|path| match fs::symlink_metadata(path) {
-        Ok(_) => true,
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => false,
-        Err(_) => true,
-    });
-    if legacy_present {
-        Config::default()
-    } else {
-        Config::fresh_install()
-    }
-}
-
-/// Pull whatever we can reuse out of the old TypeScript app's config. Favorites,
-/// history and playlists are migrated later (M5) when the library view consumes them.
-fn import_old_from(path: &std::path::Path, cfg: &mut Config) {
-    // Legacy import: cap the read and refuse a symlink, like every other persisted-state read.
-    let Ok(bytes) = crate::util::safe_fs::read_no_symlink_limited(path, MAX_CONFIG_BYTES) else {
-        return;
-    };
-    let Ok(text) = String::from_utf8(bytes) else {
-        return;
-    };
-    let Ok(json) = serde_json::from_str::<serde_json::Value>(&text) else {
-        return;
-    };
-    if let Some(v) = json.get("volume").and_then(serde_json::Value::as_i64) {
-        cfg.volume = v.clamp(0, 100);
-    }
-    if let Some(c) = json
-        .pointer("/youtubeMusic/cookie")
-        .and_then(serde_json::Value::as_str)
-        && !c.is_empty()
-    {
-        cfg.cookie = Some(c.to_owned());
-    }
-    if let Some(d) = json
-        .get("downloadDirectory")
-        .and_then(serde_json::Value::as_str)
-        && !d.is_empty()
-    {
-        cfg.download_dir = Some(PathBuf::from(d));
-    }
-}
-
-/// Turn a Netscape `cookies.txt` body into a `name=value; ...` header, keeping only
-/// youtube.com cookies (where the YTM auth cookies — SAPISID etc. — live).
-fn parse_netscape_cookies(content: &str) -> String {
-    let mut pairs = Vec::new();
-    for raw in content.lines() {
-        // `#HttpOnly_` lines are real cookies; any other leading `#` is a comment.
-        let line = raw.strip_prefix("#HttpOnly_").unwrap_or(raw);
-        if line.starts_with('#') || line.trim().is_empty() {
-            continue;
-        }
-        let fields: Vec<&str> = line.split('\t').collect();
-        if fields.len() < 7 {
-            continue;
-        }
-        // Suffix-boundary match, not `contains`: `evil-youtube.com` / `notyoutube.com` must not
-        // slip through. A Netscape domain field may carry a leading dot (`.youtube.com`).
-        let domain = fields[0].trim_start_matches('.');
-        if !(domain == "youtube.com" || domain.ends_with(".youtube.com")) {
-            continue;
-        }
-        let (name, value) = (fields[5].trim(), fields[6].trim());
-        // Drop any pair carrying header-breaking characters. A genuine cookie value never
-        // contains CR/LF/`;`; rejecting them keeps a crafted cookies.txt from injecting extra
-        // header pairs into the `name=value; …` string.
-        if name.is_empty() || name.contains(['\r', '\n', ';']) || value.contains(['\r', '\n', ';'])
-        {
-            continue;
-        }
-        pairs.push(format!("{name}={value}"));
-    }
-    pairs.join("; ")
 }
 
 #[cfg(test)]
